requirements.txt
aiofiles==23.2.1
aiogram==2.23.1
aiohttp==3.9.1
aiosignal==1.3.1
annotated-types==0.6.0
anyio==4.2.0
async-timeout==4.0.3
attrs==23.1.0
Babel==2.9.1
beautifulsoup4==4.12.2
certifi==2023.11.17
charset-normalizer==3.3.2
colorama==0.4.6
coloredlogs==15.0.1
contourpy==1.2.0
cycler==0.12.1
distro==1.8.0
fasteners==0.19
filelock==3.13.1
fonttools==4.47.0
frozenlist==1.4.1
fsspec==2023.12.2
h11==0.14.0
httpcore==1.0.2
httpx==0.25.2
huggingface-hub==0.20.1
humanfriendly==10.0
idna==3.6
kiwisolver==1.4.5
lxml==4.9.3
magic-filter==1.0.12
matplotlib==3.8.2
multidict==6.0.4
numpy==1.26.3
openai==1.5.0
packaging==23.2
pillow==10.2.0
prompt-toolkit==3.0.43
property-manager==3.0
pydantic==2.5.2
pydantic_core==2.14.5
pydub==0.25.1
pyparsing==3.1.1
pyreadline3==3.4.1
python-dateutil==2.8.2
python-dotenv==1.0.0
python-socks==2.4.4
pytz==2023.3.post1
PyYAML==6.0.1
requests==2.31.0
response==0.5.0
scipy==1.11.4
six==1.16.0
sniffio==1.3.0
soupsieve==2.5
SpeechRecognition==3.9.0
tqdm==4.66.1
typing_extensions==4.9.0
urllib3==2.1.0
verboselogs==1.7
wcwidth==0.2.12
websockets==11.0.3
wit==6.0.1
yarl==1.9.4
